fix(hardening): valores monetarios em Decimal, audit log, conciliacao e validacoes
Aplica a ultima leva de correcoes da auditoria estatica de 06/08/2026 -
os achados transversais que ficaram de fora dos clusters paralelos por
tocarem models.py e pontos de mutacao espalhados por quase todo modulo.

- A19: exportacao de "conferencia" calcula conciliacao pela igualdade
  das somas (nota+comprovante vs transacao), respeitando direcao (D/C)
  e expondo divergencia residual, em vez de "Conciliado" so por presenca
- M1: valores monetarios passam de float para Decimal em models.py,
  schemas (notas, cartoes, comprovantes, contabil, empresas, extrato,
  neo, relatorios) e nos calculos de relatorios/service.py
- M7: stats passa a filtrar Transacao.deleted_at.is_(None)
- M8: AgenciaBancaria ganha a UniqueConstraint equivalente a migration
- M9 (resto): neutraliza formula injection em exportacao/service.py
- M11 (resto): Open Banking nao devolve mais str(exc) cru ao cliente
- M12: novo modulo src/domain/auditoria com helper central de audit
  log, chamado nas mutacoes mais sensiveis (empresa, permissoes, plano
  de contas, associacao manual do NEO, fatura de cartao, conexao Open
  Banking) - nao cobre literalmente toda mutacao do sistema, escopo
  intencional
- B1: valida CNPJ, conta bancaria e competencia via tipos/validadores
  compartilhados em src/schemas/types.py

+++ src/domain/openbanking/providers/base.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from datetime import date
+from decimal import Decimal
 
 
 @dataclass
@@ -20,7 +21,7 @@ class ContaInfo:
     agencia: str | None
     numero: str | None
     tipo: str  # "CHECKING" | "SAVINGS" | "CREDIT"
-    saldo: float | None
+    saldo: Decimal | None
 
 
 @dataclass
@@ -28,7 +29,7 @@ class TransacaoInfo:
     """Dados de uma transação bancária retornados pelo provedor."""
     id_externo: str      # ID único no provedor (usado para dedup)
     data: date
-    valor: float         # sempre positivo
+    valor: Decimal       # sempre positivo
     dc: str              # "D" ou "C"
     descricao: str
     categoria: str | None
